requirements.txt
flask==2.3.3
python-dotenv==1.0.1
requests==2.32.3
retry==0.9.2
gunicorn==22.0.0